Added unit testing
